Create Signals
Create 2 Django Signals. One is used to create a Reader Profile when the
Django built-in User Model is instanciated. This allows me to create a
User object to authenticate, and a Reader object to identify the user as
a reader.

The other signal ensures that when a Reader Profile is deleted the
corresponding User account is also deleted.